`src/concepts/is_specialization_concepts.hpp`: attempt to make names without "rtargs" work by way of a peeled off `auto` template argument `src/containers/std_container_id_funcs.hpp`: support `std::unique_ptr` `src/json_stuff/json_stuff.hpp`: support `std::map` and `std::unordered_map` in serialization `src/json_stuff/json_stuff.hpp`: make `MEMB_DESERIALIZE` support `name.deserialize(...)` via a condition macro argument `game::ecs::Engine`: change type of `_num_files` to `u64` misc. other changes

+++ src/game_stuff/ecs_classes.cpp
@@ -101,13 +101,13 @@ bool Sys::_tick_helper(Engine* ecs_engine, bool cond)
 	return false;
 }
 //--------
-Engine::Engine(size_t s_num_files)
+Engine::Engine(u64 s_num_files)
 	: _next_ent_id_vec(s_num_files, 0),
 	_to_destroy_set_vec(s_num_files, EntIdSet()),
 	_num_files(s_num_files)
 	//_engine_comp_map_vec(s_num_files, EngineCompMap())
 {
-	for (size_t i=0; i<_num_files; ++i)
+	for (u64 i=0; i<_num_files; ++i)
 	{
 		_engine_comp_map_vec.push_back(EngineCompMap());
 	}
@@ -144,6 +144,10 @@ Engine::operator Json::Value () const
 	return ret;
 }
 //--------
+//void Engine::deserialize(const Json::Value& jv)
+//{
+//	MEMB_LIST_ECS_ENGINE(MEMB_DESERIALIZE);
+//}
 void Engine::_autoser_deserialize(const Json::Value& jv)
 {
 	MEMB_AUTOSER_LIST_ECS_ENGINE(MEMB_DESERIALIZE);
